+++ src/main/java/com/wanted/codebombalms/problems/set/presentation/response/ProblemSetCreateResponse.java
@@ -1,7 +1,6 @@
 package com.wanted.codebombalms.problems.set.presentation.response;
 
 import com.wanted.codebombalms.problems.set.application.command.ProblemSetCreateCommandResult;
-import com.wanted.codebombalms.problems.set.domain.model.CreatedProblemSummary;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.util.List;
@@ -40,24 +39,5 @@ public record ProblemSetCreateResponse(
                         .map(CreatedProblemResponse::from)
                         .toList()
         );
-    }
-
-    public record CreatedProblemResponse(
-            @Schema(description = "생성된 소문제 ID", example = "5001")
-            Long problemId,
-
-            @Schema(description = "소문제 순서", example = "1")
-            Integer problemOrder,
-
-            @Schema(description = "소문제 제목", example = "데이터 행과 열 개수 확인")
-            String title
-    ) {
-        private static CreatedProblemResponse from(CreatedProblemSummary problem) {
-            return new CreatedProblemResponse(
-                    problem.problemId(),
-                    problem.problemOrder(),
-                    problem.title()
-            );
-        }
     }
 }
