+++ src/main/java/com/wanted/codebombalms/problems/set/presentation/response/ProblemSetWithDatasetCreateResponse.java
@@ -67,22 +67,4 @@ public record ProblemSetWithDatasetCreateResponse(
         );
     }
 
-    public record CreatedProblemResponse(
-            @Schema(description = "생성된 소문제 ID", example = "5001")
-            Long problemId,
-
-            @Schema(description = "소문제 순서", example = "1")
-            Integer problemOrder,
-
-            @Schema(description = "소문제 제목", example = "데이터 행과 열 개수 확인")
-            String title
-    ) {
-        private static CreatedProblemResponse from(CreatedProblemSummary problem) {
-            return new CreatedProblemResponse(
-                    problem.problemId(),
-                    problem.problemOrder(),
-                    problem.title()
-            );
-        }
-    }
 }
